Move the helper to utils
